Move IMU, Datalogging, and GamePad to Lib612
