+++ src/Subsystems/Drivetrain.cpp
@@ -7,7 +7,7 @@ Drivetrain::Drivetrain(CANTalon* t_fl, CANTalon* t_fr, CANTalon* t_rl, CANTalon*
 		Subsystem("Drivetrain"),
 		RobotDrive(t_fl, t_fr, t_rl, t_rr)
 {
-	imu = new IMU(new SerialPort(57600,SerialPort::kMXP),REFRESH_RATE);
+	imu = new Lib612::IMU(new SerialPort(57600,SerialPort::kMXP),REFRESH_RATE);
 	motor_power = 1.0f;
 	fl = t_fl;
 	fr = t_fr;
